2793 erro excluir selecao com

+++ aplicacoes/gpa-mae/src/main/java/br/ufc/quixada/npi/gpa/controller/CoordenadorController.java
@@ -233,13 +233,13 @@ public class CoordenadorController {
 
 		Selecao selecao = this.selecaoService.find(Selecao.class, idSelecao);
 
-		if (selecao != null) {
+		if (selecao != null && selecao.getInscritos().isEmpty()) {
 
 			this.selecaoService.delete(selecao);
 			redirect.addFlashAttribute("info", MENSAGEM_SUCESSO_SELECAO_REMOVIDA);
 
 		} else {
-			redirect.addFlashAttribute("erro", MENSAGEM_ERRO_SELECAO_INEXISTENTE);
+			redirect.addFlashAttribute("erro", MENSAGEM_ERRO_EXCLUIR_SELECAO_COM_INSCRITOS);
 
 		}
